refactor: modal type separation
