+++ my-monorepo/packages/ui/.eslintrc.js
@@ -1,10 +1,20 @@
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
-  extends: ["@repo/eslint-config/react-internal.js", "plugin:storybook/recommended"],
+  extends: [
+    "@repo/eslint-config/react-internal.js",
+    "plugin:storybook/recommended",
+  ],
   parser: "@typescript-eslint/parser",
   parserOptions: {
     project: "./tsconfig.lint.json",
     tsconfigRootDir: __dirname,
+  },
+  rules: {
+    ...reactHooks.configs.recommended.rules,
+    "react-refresh/only-export-components": [
+      "warn",
+      { allowConstantExport: true },
+    ],
   },
 };
